[DROOLS-6105] fix query corner cases in executable model
