+++ drools-model/drools-canonical-model/src/main/java/org/drools/model/AccumulatePattern.java
@@ -25,6 +25,8 @@ public interface AccumulatePattern<T> extends Pattern<T> {
 
     boolean isCompositePatterns();
 
+    boolean isQuerySource();
+
     Pattern getPattern();
 
     Condition getCondition();
