+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/builder/generator/declaredtype/generator/GeneratedConstructor.java
@@ -110,9 +110,9 @@ class FullArgumentConstructor implements GeneratedConstructor {
             Type returnType = parseType(fieldDefinition.getObjectType());
             addConstructorArgument(constructor, returnType, fieldName);
             fieldAssignStatement.add(fieldAssignment(fieldName));
-
-            constructor.setBody(new BlockStmt(fieldAssignStatement));
         }
+
+        constructor.setBody(new BlockStmt(fieldAssignStatement));
 
         if (constructor.getParameters().isNonEmpty()) {
             generatedClass.addMember( constructor );
