+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/constraints/UnificationConstraint.java
@@ -34,24 +34,24 @@ import org.drools.modelcompiler.constraints.LambdaConstraint.LambdaContextEntry;
 
 public class UnificationConstraint extends MutableTypeConstraint implements IndexableConstraint {
 
-    private final Declaration declaration;
+    private Declaration indexingDeclaration;
     private final InternalReadAccessor readAccessor;
     private final ConstraintEvaluator evaluator;
 
     private boolean unification = true;
 
-    public UnificationConstraint( Declaration declaration ) {
-        this(declaration, null);
+    public UnificationConstraint( Declaration indexingDeclaration ) {
+        this( indexingDeclaration, null);
     }
 
-    public UnificationConstraint( Declaration declaration, ConstraintEvaluator evaluator ) {
-        this.declaration = declaration;
+    public UnificationConstraint( Declaration indexingDeclaration, ConstraintEvaluator evaluator ) {
+        this.indexingDeclaration = indexingDeclaration;
         this.evaluator = evaluator;
         if (evaluator != null) {
             Index index = evaluator.getIndex();
             this.readAccessor = new LambdaReadAccessor( index.getIndexId(), index.getIndexedClass(), index.getLeftOperandExtractor() );
         } else {
-            this.readAccessor = new LambdaReadAccessor( declaration.getDeclarationClass(), x -> x );
+            this.readAccessor = new LambdaReadAccessor( indexingDeclaration.getDeclarationClass(), x -> x );
         }
     }
 
@@ -82,7 +82,7 @@ public class UnificationConstraint extends MutableTypeConstraint implements Inde
 
     @Override
     public FieldIndex getFieldIndex() {
-        return new FieldIndex(readAccessor, declaration);
+        return new FieldIndex(readAccessor, indexingDeclaration );
     }
 
     @Override
@@ -92,26 +92,27 @@ public class UnificationConstraint extends MutableTypeConstraint implements Inde
 
     @Override
     public Declaration getIndexExtractor() {
-        return declaration;
+        return indexingDeclaration;
     }
 
     @Override
     public Declaration[] getRequiredDeclarations() {
-        return new Declaration[] { declaration };
+        return new Declaration[] { indexingDeclaration };
     }
 
     @Override
     public void replaceDeclaration( Declaration oldDecl, Declaration newDecl ) {
+        if (indexingDeclaration == oldDecl) {
+            indexingDeclaration = newDecl;
+        }
         if (evaluator != null) {
             evaluator.replaceDeclaration( oldDecl, newDecl );
-        } else {
-            throw new UnsupportedOperationException();
         }
     }
 
     @Override
     public MutableTypeConstraint clone() {
-        return new UnificationConstraint( declaration, evaluator );
+        return new UnificationConstraint( indexingDeclaration, evaluator );
     }
 
     @Override
@@ -139,13 +140,13 @@ public class UnificationConstraint extends MutableTypeConstraint implements Inde
             return evaluator.evaluate(handle, tuple, workingMemory);
         }
         DroolsQuery query = ( DroolsQuery ) tuple.getObject( 0 );
-        if (query.getVariables()[declaration.getExtractor().getIndex()] != null) {
+        if (query.getVariables()[indexingDeclaration.getExtractor().getIndex()] != null) {
             return true;
         }
         if (evaluator != null) {
             return evaluator.evaluate(handle, tuple, workingMemory);
         }
-        Object argument = declaration.getValue( null, query );
+        Object argument = indexingDeclaration.getValue( null, query );
         return handle.getObject().equals( argument );
     }
